feature: Getting ready to publish the project
(cherry picked from commit 64ff5ad5f6b34ec54fcf1194a4d7a72d6c0e9b20)

+++ site/src/commonMain/kotlin/org/example/blogmultiplatform/models/Post.kt
@@ -9,7 +9,7 @@ data class Post(
     @SerialName("_id")
     val id: String = "",
     val author: String = "",
-    val date: Long = 0L,
+    val date: Double = 0.0,
     val title: String,
     val subtitle: String,
     val thumbnail: String,
@@ -25,7 +25,7 @@ data class SimplePost(
     @SerialName("_id")
     val id: String = "",
     val author: String,
-    val date: Long,
+    val date: Double,
     val title: String,
     val subtitle: String,
     val thumbnail: String,
